requirements.txt
attrs==21.4.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
dash==2.1.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Fiona==1.8.21
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.29.1
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
munch==2.5.0
mysql-connector==2.2.9
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
PyMySQL==1.0.2
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
Shapely==1.8.0
six==1.16.0
SQLAlchemy==1.4.31
tenacity==8.0.1
tweepy==4.5.0
urllib3==1.26.8
Werkzeug==2.0.3
